fix ci

+++ elasticdl/pkg/kernel/capi/kernel_api.cc
@@ -4,7 +4,7 @@ extern "C" {
 
 #include <eigen3/Eigen/Dense>
 
-void SGD(float* grad, float* param, double lr, int64_t size) {
+void SGD(float* grad, float* param, double lr, long long size) {
   Eigen::Map<Eigen::Array<float, 1, Eigen::Dynamic>> eg{
       grad, static_cast<Eigen::Index>(size)};
